tests: rename class
